requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
seaborn
matplotlib
scikit-learn
tensorflow

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
s3fs
fastapi
uvicorn

# utilities
six>=1.14
joblib
memoized-property
termcolor

# specific
tldextract
nltk==3.6.2
bs4

tensorflow_text
tf-models-official
tensorflow-addons
tensorflow-model-optimization>=0.1.3

